# perf: using asynchronous worker to validate BLS signatures in quorum commitments