--- a/src/llmq/blockprocessor.cpp
+++ b/src/llmq/blockprocessor.cpp
@@ -66,7 +66,7 @@ CQuorumBlockProcessor::CQuorumBlockProcessor(CChainState& chainstate, CDetermini
     // Number of script-checking threads <= MAX_BLSCHECK_THREADS
     bls_threads = std::min(bls_threads, MAX_BLSCHECK_THREADS);
 
-    LogPrintf("Bls verification uses %d additional threads\n", bls_threads);
+    LogPrintf("BLS verification uses %d additional threads\n", bls_threads);
     m_bls_queue.StartWorkerThreads(bls_threads);
 }