--- a/src/llmq/commitment.h
+++ b/src/llmq/commitment.h
@@ -25,11 +25,16 @@ class CBlockIndex;
 class CDeterministicMNManager;
 class ChainstateManager;
 class TxValidationState;
+template <typename T>
+class CCheckQueueControl;
 
 namespace llmq
 {
 class CQuorumSnapshotManager;
 
+namespace utils {
+struct BlsCheck;
+} // namespace utils
 // This message is an aggregation of all received premature commitments and only valid if
 // enough (>=threshold) premature commitments were aggregated
 // This is mined on-chain as part of TRANSACTION_QUORUM_COMMITMENT
@@ -67,8 +72,9 @@ public:
         return int(std::count(validMembers.begin(), validMembers.end(), true));
     }
 
-    bool VerifySignature(CDeterministicMNManager& dmnman, CQuorumSnapshotManager& qsnapman,
-                         gsl::not_null<const CBlockIndex*> pQuorumBaseBlockIndex) const;
+    bool VerifySignatureAsync(CDeterministicMNManager& dmnman, CQuorumSnapshotManager& qsnapman,
+                              gsl::not_null<const CBlockIndex*> pQuorumBaseBlockIndex,
+                              CCheckQueueControl<utils::BlsCheck>* queue_control) const;
     bool Verify(CDeterministicMNManager& dmnman, CQuorumSnapshotManager& qsnapman,
                 gsl::not_null<const CBlockIndex*> pQuorumBaseBlockIndex, bool checkSigs) const;
     bool VerifyNull() const;