@knst 2025-06-17
> @[knst](https://github.com/knst) knst [force-pushed](https://github.com/dashpay/dash/compare/8b6414d246c66a6ea4a82e466a4f3e15f4a97c0f..da36faf5c0b41c6241d316acc22538fa2bfdc511) the perf-bls-parallel branch from [8b6414d](https://github.com/dashpay/dash/commit/8b6414d246c66a6ea4a82e466a4f3e15f4a97c0f) to [da36faf](https://github.com/dashpay/dash/commit/da36faf5c0b41c6241d316acc22538fa2bfdc511) [June 17, 2025 16:06](https://github.com/dashpay/dash/pull/6692#event-18186416245)

I squashed commits together because commits has changed same piece of code several times and moved between files.

Also see https://github.com/dashpay/dash/pull/6724 and https://github.com/dashpay/dash/pull/6692#issuecomment-2916034041 for extra tests done for this PR.

@UdjinM6 please help with review, PR is ready now (assuming CI will succeed)