@knst 2025-08-19
> @[knst](https://github.com/knst) knst [force-pushed](https://github.com/dashpay/dash/compare/6906aa223d1830b1abf8f6d589c215d61211e28e..7999733d0078cd852e31bd99fd826fcfa1adb593) the perf-bls-parallel branch from [6906aa2](https://github.com/dashpay/dash/commit/6906aa223d1830b1abf8f6d589c215d61211e28e) to [7999733](https://github.com/dashpay/dash/commit/7999733d0078cd852e31bd99fd826fcfa1adb593) [August 19, 2025 14:28](https://github.com/dashpay/dash/pull/6692#event-19218384292)

Force-pushed to resolve conflict with `develop`